Ajustar stock estaba roto desde el lunes, y lo rompí yo
`could not determine data type of parameter $5`. Un parámetro que sólo aparece
adentro de `IS NOT NULL` no le dice a Postgres de qué tipo es, y Postgres no
adivina: rechaza la consulta entera.

Entró con los márgenes, cuando agregué `precio_manual` para distinguir el
precio que calculó una regla del que escribió el comerciante. Cambié cuatro
consultas del camino de la plata y sólo corrí el typecheck. TypeScript no mira
adentro de un string de SQL, así que compiló perfecto y se rompió en producción.

Rotas desde el 16:

  POST /api/stock/adjust            ← el que reportó German
  POST /api/stock/add-from-catalog  ← nadie lo reportó todavía

Las otras dos del mismo patrón NO estaban rotas, y vale la pena saber por qué:
en ellas el mismo parámetro también aparece dentro de un COALESCE contra una
columna, y eso alcanza para que Postgres le ponga tipo. O sea que el patrón no
falla siempre, que es lo que lo hace fácil de dejar pasar.

Comprobado contra la base, con PREPARE, cuál acepta y cuál no:

  stock sin cast            → rechazada
  stock con ::numeric       → PREPARE
  clubpay sin cast          → PREPARE   (el COALESCE le da el tipo)
  /precio sin cast          → PREPARE   (ídem)

Arreglados los cuatro igual: que uno funcione por un COALESCE que está al lado
es una casualidad, y la próxima edición se lo puede llevar puesto.

Verificado en los tres caminos del ajuste —sólo cantidad, con precio y mínimo, y
un producto que no estaba en el stock— más el alta desde catálogo con y sin
precio, y la celda de precio de la tabla.

+++ backend/src/modules/stock.ts
@@ -219,7 +219,7 @@ stockRouter.put("/:productId/precio", async (req, res, next) => {
        SET sale_price = COALESCE($1, sale_price),
            -- Un precio que escribió el comerciante queda suyo: la próxima
            -- corrida de márgenes lo saltea en vez de pisárselo.
-           precio_manual = (stock_items.precio_manual OR $1 IS NOT NULL),
+           precio_manual = (stock_items.precio_manual OR $1::numeric IS NOT NULL),
            cost = COALESCE($2, cost),
            min_stock = COALESCE($3, min_stock),
            updated_at = now()
@@ -257,12 +257,12 @@ stockRouter.post("/adjust", async (req, res, next) => {
     } = await client.query(
       `INSERT INTO stock_items (commerce_id, product_id, quantity, cost, sale_price,
                                 min_stock, precio_manual, updated_at)
-       VALUES ($1, $2, GREATEST($3, 0), $4, $5, COALESCE($6, 0), $5 IS NOT NULL, now())
+       VALUES ($1, $2, GREATEST($3, 0), $4, $5, COALESCE($6, 0), $5::numeric IS NOT NULL, now())
        ON CONFLICT (commerce_id, product_id) DO UPDATE SET
          quantity = GREATEST(stock_items.quantity + $3, 0),
          cost = COALESCE($4, stock_items.cost),
          sale_price = COALESCE($5, stock_items.sale_price),
-         precio_manual = (stock_items.precio_manual OR $5 IS NOT NULL),
+         precio_manual = (stock_items.precio_manual OR $5::numeric IS NOT NULL),
          min_stock = COALESCE($6, stock_items.min_stock),
          updated_at = now()
        RETURNING id, quantity`,
@@ -376,12 +376,12 @@ stockRouter.post("/add-from-catalog", async (req, res, next) => {
     await client.query(
       `INSERT INTO stock_items (commerce_id, product_id, quantity, cost, sale_price,
                                 min_stock, precio_manual, updated_at)
-       VALUES ($1, $2, $3, $4, $5, COALESCE($6, 0), $5 IS NOT NULL, now())
+       VALUES ($1, $2, $3, $4, $5, COALESCE($6, 0), $5::numeric IS NOT NULL, now())
        ON CONFLICT (commerce_id, product_id) DO UPDATE SET
          quantity = stock_items.quantity + EXCLUDED.quantity,
          cost = COALESCE($4, stock_items.cost),
          sale_price = COALESCE($5, stock_items.sale_price),
-         precio_manual = (stock_items.precio_manual OR $5 IS NOT NULL),
+         precio_manual = (stock_items.precio_manual OR $5::numeric IS NOT NULL),
          min_stock = COALESCE($6, stock_items.min_stock),
          updated_at = now()`,
       [commerceId, product.id, body.quantity, body.cost ?? null, body.salePrice ?? null, body.minStock ?? null]
